backend/requirements.txt
Flask==3.0.3
waitress==3.0.0
pylint==3.2.2
yapf==0.3.1
flask-marshmallow==1.2.1
SQLAlchemy==2.0.30
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.9
marshmallow-sqlalchemy==1.0.0
Flask-Cors==4.0.1
redis==5.2.1